requirements.txt
flask
line-bot-sdk
opencv-python-headless
requests
